+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ---------
 
 v1.12.0 (??.0?.2020)
+ * raised required PHP version to 7.1
  * added support for TYPO3 10.4 LTS
  * dropped support for TYPO3 6.2 LTS
  * #165 wip refactoring of classes to PSR-4